requirements.txt
# Apache Airflow and providers
apache-airflow>=2.5.0
apache-airflow-providers-google>=8.0.0
apache-airflow-providers-docker>=3.0.0

# Data processing
numpy>=1.20.0
pandas>=1.4.0
polars>=0.17.0
rapidfuzz>=2.0.0
scikit-learn>=1.0.0

# Cloud storage
google-cloud-storage>=2.0.0
fsspec>=2022.1.0

# Data validation
tensorflow-data-validation>=1.10.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
pyarrow>=10.0.0
openpyxl>=3.0.0
xlrd>=2.0.0

# For testing
pytest>=7.0.0

# Email functionality
sendgrid>=6.0.0  # If using SendGrid for emails 